Made Etch-A-Sketch program for Odin Project
User can submit an integer and the sketchpad will have that number of
rows and columns. If you hover the mouse over the square sketchpad,
pixels will turn black. Since the integer user typed in dictates the
number of pixels the sketchpad will have, it is better to type in higher
integers for better resolution. I wanted to implement 'click' event to
turn pen on and off but decided that it was going to take more time
than I can afford right now. I will learn how to click on and off by
reviewing what others have done. For now, I am just satisfied that I was
able to finish the project that fulfills the Odin Project's parameters.